Category

+++ src/FCG.Catalog.Domain/Entities/Game.cs
@@ -2,10 +2,10 @@
 
 public class Game
 {
-    public int Id { get; set; }
+    public long Id { get; set; }
     public string Name { get; set; } = string.Empty;
     public string Description { get; set; } = string.Empty;
     public decimal Price { get; set; }
-    public int CategoryId { get; set; }
+    public long CategoryId { get; set; }
     public Category Category { get; set; } = default!;
 }
